Add per-runner memory limits and fix worker bugs
- lxc.js: pass memLimit to lxc-start-ephemeral, add setMemLimit writing to
  the live cgroup v2 path, add parseMemLimit helper
- codeland.js: accept memLimit config default and per-request override,
  fix oven >= check, remove debug log, clean redundant check
- routes/runner.js: accept memLimit on /run and /new
- conf: add memLimit default, fix production domian typo
- Fix worker.js next(err) and deleteUntrackedRunners typo
- Fix group_ldap.js require path, redis_model exists(), ssh error guard,
  object_validate string max message

+++ nodejs/lib/object_validate.js
@@ -7,7 +7,7 @@ const process_type = {
 	},
 	string: function(key, value){
 		if(key.min && value.length < key.min) return `is too short, min ${key.min}.`
-		if(key.max && value.length > key.max) return `is too short, max ${key.max}.`
+		if(key.max && value.length > key.max) return `is too long, max ${key.max}.`
 	},
 }
